+++ NEWS.md
@@ -1,5 +1,8 @@
 # httr2 (development version)
 
+* New `req_paginate()` and `paginate_req_perform()` to easily request paginated
+  APIs (@mgirlich, #8).
+
 * New `resp_check_content_type()` to check response content types (#190).
   `resp_body_json()` and friends give better errors if no `Content-Type`
   header is present in the response (#284).